Split library and sample in two modules
