+++ build.gradle
@@ -1,5 +1,4 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
     repositories {
         jcenter()
@@ -7,6 +6,19 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:1.1.3'
         classpath 'com.github.dcendents:android-maven-plugin:1.2'
-        
     }
 }
+
+ext {
+    compileSdkVersion = 21
+    buildToolsVersion = '22.0.0'
+    minSdkVersion = 15
+    targetSdkVersion = 21
+}
+
+///////////////////////////////////
+// Misc dependencies and properties
+///////////////////////////////////
+
+// System properties
+ext.preDexLibs = !project.hasProperty('disablePreDex')
