+++ library/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="io.techery.snapper">
+
+    <application />
+
+</manifest>
