+++ library/src/main/java/io/techery/snapper/DataCollection.java
