+++ library/src/main/java/io/techery/snapper/DroidSnapper.java
