+++ library/src/main/java/io/techery/snapper/model/Indexable.java
