+++ library/src/main/java/io/techery/snapper/snappydb/SnappyDBAdapter.java
