+++ library/src/main/java/io/techery/snapper/storage/DatabaseAdapter.java
