+++ library/src/main/java/io/techery/snapper/storage/PersistentStorage.java
