+++ library/src/main/java/io/techery/snapper/storage/Storage.java
