+++ library/src/main/java/io/techery/snapper/storage/StorageChange.java
