+++ library/src/main/java/io/techery/snapper/view/IDataView.java
