+++ sample/src/main/AndroidManifest.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="techery.io.snappytest"
+    package="com.example.snapper"
     >
 
     <application
-        android:name=".App"
+        android:name="com.example.snapper.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme"
         >
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.snapper.MainActivity"
             android:label="@string/app_name"
             >
             <intent-filter>
