+++ sample/src/main/res/layout/user_cell.xml
